Update main.yml
Install gitconfig task copies a pre-populated gitconfig. This is a leftover from inherited code and not used in TNT as of today. 

It should be reworked into a native ansible templating task later on. Since it's not currently needed, we're just letting it error out when the user's ansible-running computer doesn't have the right Python dependencies